+++ ChangeLog.md
@@ -3,6 +3,8 @@ AWS Lambda Webservices change log
 
 ## ?.?.? / ????-??-??
 
+* Fixed `FromApiGateway::readLine()` not returning EOF correctly - @thekid
+
 ## 0.4.0 / 2021-08-29
 
 * Fixed issue #2: Return cookies in multiValueHeaders - @thekid